File - package system reworked.
