+++ Infrastructure/Sources/CoreToolkit/String+.swift
