+++ Infrastructure/Sources/Networking/URLs.swift
@@ -9,9 +9,9 @@ import Foundation
 public struct Constants {
     enum Networking {
         static let baseURL = "https://api.spacexdata.com"
-        static let allRockets = "/v4/rockets"
+        static let allRockets = baseURL + "/v4/rockets"
         static func oneRocket(id: String) -> String {
-            return "/v4/rockets/" + id
+            return baseURL + "/v4/rockets/" + id
         }
     }
 }
