+++ Infrastructure/Sources/RocketModels/RocketDetailModel.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by David Jilek on 20.09.2022.
+//
+
+import Foundation
