--- Infrastructure/Tests/ObjectModelTests/ObjectModelTests.swift
@@ -1,5 +1,0 @@
-import XCTest
-@testable import ObjectModel
-
-final class NetworkingTests: XCTestCase {
-}
